Java advanced/repo lkr/1 Lambda and streams/ex2/src/main/java/fact/it/ex2/model/Student.java
package fact.it.ex2.model;

import java.util.List;

public class Student {
    private String firstName;
    private String lastName;
    private Integer age;
    private List<Course> courses;

    public Student() {
    }

    public Student(String firstName, String lastName, Integer age, List<Course> courses) {
        this.firstName = firstName;
        this.lastName = lastName;
        this.age = age;
        this.courses = courses;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public Integer getAge() {
        return age;
    }

    public void setAge(Integer age) {
        this.age = age;
    }

    public List<Course> getCourses() {
        return courses;
    }

    public void setCourses(List<Course> courses) {
        this.courses = courses;
    }
}
